pyproject.toml
[tool.poetry]
name = "ssh-key-server"
version = "0.1.0"
description = "A simple tool to generate SSH keys"
authors = ["Morgan Gangwere <morgan.gangwere@gmail.com>"]
license = "BSD-1-Clause"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.0.1"
pyotp = "^2.6.0"
pycryptodome = "^3.10.1"
python-pam = "^1.8.4"
tinydb = "^4.5.1"
qrcode = {extras = ["pil"], version = "^7.3"}
python-dotenv = "^0.19.0"
Flask-Limiter = "^1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
